--- src/main/java/nbbang/com/nbbang/bbangpan/controller/BbangpanController.java
@@ -1,4 +1,0 @@
-package nbbang.com.nbbang.bbangpan.controller;
-
-public class BbangpanController {
-}
