--- src/main/java/nbbang/com/nbbang/bbangpan/dto/BbangpanDto.java
@@ -1,4 +1,0 @@
-package nbbang.com.nbbang.bbangpan.dto;
-
-public class BbangpanDto {
-}
