--- src/main/java/nbbang/com/nbbang/bbangpan/repository/MemberPartyRepository.java
@@ -1,7 +1,0 @@
-package nbbang.com.nbbang.bbangpan.repository;
-
-import nbbang.com.nbbang.bbangpan.domain.MemberParty;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MemberPartyRepository extends JpaRepository<MemberParty, Long> {
-}
