--- src/main/java/nbbang/com/nbbang/bbangpan/service/BbangpanService.java
@@ -1,4 +1,0 @@
-package nbbang.com.nbbang.bbangpan.service;
-
-public class BbangpanService {
-}
