--- src/main/java/nbbang/com/nbbang/chat/controller/ChatController.java
@@ -1,4 +1,0 @@
-package nbbang.com.nbbang.chat.controller;
-
-public class ChatController {
-}
