--- src/main/java/nbbang/com/nbbang/chat/dto/ChatDto.java
@@ -1,4 +1,0 @@
-package nbbang.com.nbbang.chat.dto;
-
-public class ChatDto {
-}
