--- src/main/java/nbbang/com/nbbang/chat/repository/MessageRepository.java
@@ -1,7 +1,0 @@
-package nbbang.com.nbbang.chat.repository;
-
-import nbbang.com.nbbang.chat.domain.Message;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MessageRepository extends JpaRepository<Message, Long> {
-}
