--- src/main/java/nbbang/com/nbbang/chat/service/ChatService.java
@@ -1,4 +1,0 @@
-package nbbang.com.nbbang.chat.service;
-
-public class ChatService {
-}
