+++ src/main/java/nbbang/com/nbbang/domain/bbangpan/controller/BbangpanController.java
@@ -1,0 +1,4 @@
+package nbbang.com.nbbang.domain.bbangpan.controller;
+
+public class BbangpanController {
+}
