+++ src/main/java/nbbang/com/nbbang/domain/bbangpan/dto/BbangpanDto.java
@@ -1,0 +1,4 @@
+package nbbang.com.nbbang.domain.bbangpan.dto;
+
+public class BbangpanDto {
+}
