+++ src/main/java/nbbang/com/nbbang/domain/bbangpan/repository/MemberPartyRepository.java
@@ -1,0 +1,7 @@
+package nbbang.com.nbbang.domain.bbangpan.repository;
+
+import nbbang.com.nbbang.domain.bbangpan.domain.MemberParty;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MemberPartyRepository extends JpaRepository<MemberParty, Long> {
+}
