+++ src/main/java/nbbang/com/nbbang/domain/bbangpan/service/BbangpanService.java
@@ -1,0 +1,4 @@
+package nbbang.com.nbbang.domain.bbangpan.service;
+
+public class BbangpanService {
+}
