+++ src/main/java/nbbang/com/nbbang/domain/chat/controller/ChatController.java
@@ -1,0 +1,4 @@
+package nbbang.com.nbbang.domain.chat.controller;
+
+public class ChatController {
+}
