+++ src/main/java/nbbang/com/nbbang/domain/chat/dto/ChatDto.java
@@ -1,0 +1,4 @@
+package nbbang.com.nbbang.domain.chat.dto;
+
+public class ChatDto {
+}
