+++ src/main/java/nbbang/com/nbbang/domain/chat/repository/MessageRepository.java
@@ -1,0 +1,7 @@
+package nbbang.com.nbbang.domain.chat.repository;
+
+import nbbang.com.nbbang.domain.chat.domain.Message;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MessageRepository extends JpaRepository<Message, Long> {
+}
