+++ src/main/java/nbbang/com/nbbang/domain/chat/service/ChatService.java
@@ -1,0 +1,4 @@
+package nbbang.com.nbbang.domain.chat.service;
+
+public class ChatService {
+}
