+++ src/main/java/nbbang/com/nbbang/domain/member/controller/MemberController.java
@@ -1,0 +1,8 @@
+package nbbang.com.nbbang.domain.member.controller;
+
+import org.springframework.stereotype.Controller;
+
+@Controller
+public class MemberController {
+    // Test purpose controller
+}
