+++ src/main/java/nbbang/com/nbbang/domain/member/dto/MemberDto.java
@@ -1,0 +1,4 @@
+package nbbang.com.nbbang.domain.member.dto;
+
+public class MemberDto {
+}
