+++ src/main/java/nbbang/com/nbbang/domain/member/repository/MemberRepository.java
@@ -1,0 +1,7 @@
+package nbbang.com.nbbang.domain.member.repository;
+
+import nbbang.com.nbbang.domain.member.domain.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MemberRepository extends JpaRepository<Member, Long> {
+}
