+++ src/main/java/nbbang/com/nbbang/domain/member/service/MemberService.java
@@ -1,0 +1,4 @@
+package nbbang.com.nbbang.domain.member.service;
+
+public class MemberService {
+}
