+++ src/main/java/nbbang/com/nbbang/domain/party/domain/Status.java
@@ -1,0 +1,7 @@
+package nbbang.com.nbbang.domain.party.domain;
+
+/*모집 중, 마감 임박, 모집 완료, 주문 완료, 모집 취소*/
+
+public enum Status {
+    ON, SOON, FINISH, ORDER, CANCEL
+}
