+++ src/main/java/nbbang/com/nbbang/domain/party/dto/PartyDto.java
@@ -1,0 +1,4 @@
+package nbbang.com.nbbang.domain.party.dto;
+
+public class PartyDto {
+}
