+++ src/main/java/nbbang/com/nbbang/domain/party/repository/HashtagRepository.java
@@ -1,0 +1,7 @@
+package nbbang.com.nbbang.domain.party.repository;
+
+import nbbang.com.nbbang.domain.party.domain.Hashtag;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface HashtagRepository extends JpaRepository<Hashtag, Long> {
+}
