+++ src/main/java/nbbang/com/nbbang/domain/party/repository/PartyRepository.java
@@ -1,0 +1,7 @@
+package nbbang.com.nbbang.domain.party.repository;
+
+import nbbang.com.nbbang.domain.party.domain.Party;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PartyRepository extends JpaRepository<Party, Long> {
+}
