+++ src/main/java/nbbang/com/nbbang/domain/party/service/PartyService.java
@@ -1,0 +1,4 @@
+package nbbang.com.nbbang.domain.party.service;
+
+public class PartyService {
+}
