+++ src/main/java/nbbang/com/nbbang/domain/web/controller/WebController.java
@@ -1,0 +1,4 @@
+package nbbang.com.nbbang.domain.web.controller;
+
+public class WebController {
+}
