+++ src/main/java/nbbang/com/nbbang/domain/web/domain/WebDomain.java
@@ -1,0 +1,4 @@
+package nbbang.com.nbbang.domain.web.domain;
+
+public class WebDomain {
+}
