+++ src/main/java/nbbang/com/nbbang/domain/web/dto/WebDto.java
@@ -1,0 +1,4 @@
+package nbbang.com.nbbang.domain.web.dto;
+
+public class WebDto {
+}
