+++ src/main/java/nbbang/com/nbbang/domain/web/repository/WebRepository.java
@@ -1,0 +1,4 @@
+package nbbang.com.nbbang.domain.web.repository;
+
+public class WebRepository {
+}
