+++ src/main/java/nbbang/com/nbbang/domain/web/service/WebService.java
@@ -1,0 +1,4 @@
+package nbbang.com.nbbang.domain.web.service;
+
+public class WebService {
+}
