+++ src/main/java/nbbang/com/nbbang/global/exception/HelloException.java
@@ -1,0 +1,4 @@
+package nbbang.com.nbbang.global.exception;
+
+public class HelloException {
+}
