--- src/main/java/nbbang/com/nbbang/member/controller/MemberController.java
@@ -1,8 +1,0 @@
-package nbbang.com.nbbang.member.controller;
-
-import org.springframework.stereotype.Controller;
-
-@Controller
-public class MemberController {
-    // Test purpose controller
-}
