--- src/main/java/nbbang/com/nbbang/member/dto/MemberDto.java
@@ -1,4 +1,0 @@
-package nbbang.com.nbbang.member.dto;
-
-public class MemberDto {
-}
