--- src/main/java/nbbang/com/nbbang/member/service/MemberService.java
@@ -1,4 +1,0 @@
-package nbbang.com.nbbang.member.service;
-
-public class MemberService {
-}
