--- src/main/java/nbbang/com/nbbang/party/domain/Party.java
@@ -1,38 +1,0 @@
-package nbbang.com.nbbang.party.domain;
-
-import lombok.Getter;
-import nbbang.com.nbbang.member.domain.Member;
-
-import javax.persistence.*;
-import java.time.LocalDateTime;
-
-@Entity @Getter
-public class Party {
-    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "party_id")
-    private Long id;
-
-    private String title;
-
-    @Column(columnDefinition = "TEXT")
-    private String content;
-
-    private LocalDateTime createTime;
-
-    private Integer goalNumber;
-
-    private String partyStatus;
-
-    private String place;
-
-    private LocalDateTime cancelTime;
-
-    private Integer fee;
-
-    @ManyToOne
-    @JoinColumn(name = "member_id")
-    private Member owner;
-
-    private Boolean isBlocked;
-
-}
