--- src/main/java/nbbang/com/nbbang/party/dto/PartyDto.java
@@ -1,4 +1,0 @@
-package nbbang.com.nbbang.party.dto;
-
-public class PartyDto {
-}
