--- src/main/java/nbbang/com/nbbang/party/service/PartyService.java
@@ -1,4 +1,0 @@
-package nbbang.com.nbbang.party.service;
-
-public class PartyService {
-}
