--- src/main/java/nbbang/com/nbbang/web/controller/WebController.java
@@ -1,4 +1,0 @@
-package nbbang.com.nbbang.web.controller;
-
-public class WebController {
-}
