--- src/main/java/nbbang/com/nbbang/web/domain/WebDomain.java
@@ -1,4 +1,0 @@
-package nbbang.com.nbbang.web.domain;
-
-public class WebDomain {
-}
