--- src/main/java/nbbang/com/nbbang/web/dto/WebDto.java
@@ -1,4 +1,0 @@
-package nbbang.com.nbbang.web.dto;
-
-public class WebDto {
-}
