--- src/main/java/nbbang/com/nbbang/web/repository/WebRepository.java
@@ -1,4 +1,0 @@
-package nbbang.com.nbbang.web.repository;
-
-public class WebRepository {
-}
