--- src/main/java/nbbang/com/nbbang/web/service/WebService.java
@@ -1,4 +1,0 @@
-package nbbang.com.nbbang.web.service;
-
-public class WebService {
-}
